\name{bikes}
\alias{bikes}
\alias{bikes}
\docType{package}
\title{
Capital Bikeshare ride data
A data frame with 848,755 rows and 7 columns.
}
\description{
This dataset contains data on all bike trips taken in July, August, and
September of 2013 on Washington, DC's bikeshare program. The data is cleaned and
some of the original variables were removed to reduce the data size.
\url{http://www.capitalbikeshare.com/system-data} The cleaning scripts are located
here: \url{https://github.com/ahalterman/bikeshare}

This work was funded as part of DARPA's XDATA program.
}
}
\details{
\tabular{ll}{
Package: \tab bikes\cr
Type: \tab Package\cr
Version: \tab 0.1\cr
Date: \tab 2014-09-11\cr
License: \tab MIT\cr
}
This is a data-only package that includes a 848,755 row, 7 column dataset.
}
\author{
Andy Halterman

Maintainer: Andy Halterman <ahalterman0@gmail.com>
}
\examples{
head(bikes)
str(bikes)
}
